frontend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
numpy==1.21.0
Werkzeug==3.0.1
python-dotenv==0.19.0
tensorflow==2.8.0
rasterio==1.2.10
shapely==1.8.0
geopandas==0.10.2
requests==2.26.0
gunicorn==20.1.0
opencv-python==4.5.3.56
pillow==8.3.2 